hot-fix: dday를 구할 때 Period.between() 대신 ChronoUnit.DAYS.between() 사용

+++ src/main/kotlin/com/yourssu/balanssu/domain/model/dto/CategoryDto.kt
@@ -3,7 +3,7 @@ package com.yourssu.balanssu.domain.model.dto
 class CategoryDto(
     val categoryId: String,
     val title: String,
-    val dDay: Int,
+    val dDay: Long,
     val participantCount: Int,
     val isParticipating: Boolean,
     val myChoice: String?
